+++ R/repositories-releases.R
@@ -68,7 +68,7 @@ view_releases <- function(
       error("'api' must be a valid URL:\n  '", paste(api, collapse = "'\n  '"), "'")
   }
 
-  if (missing(tags) || is_null(tags) || is_na(tags)) {
+  if (missing(tags) || is_null(tags) || all(is_na(tags))) {
     info("Getting up to ", n_max, " releases from repository '", repo, "'")
 
     releases_list <- try_catch({