+++ tests/testthat/test-github-api.R
@@ -99,54 +99,56 @@ test_that("gh_url returns a valid URL for the GitHub API", {
 
 # TEST: gh_request ----------------------------------------------------------------------------
 
+master_sha <- view_shas(refs = "master", repo = "ChadGoymer/test-githapi")[[1]]
+test_sha <- view_shas(refs = "test-files", repo = "ChadGoymer/test-githapi")[[1]]
+
 test_that("gh_request can GET, POST and DELETE a tag in the specified repository", {
 
+  test_tag <- str_c("refs/tags/test-gh-request-", format(Sys.time(), "%Y-%m-%d-%H-%M-%S"))
   created_tag <- gh_request(
-    url     = "https://api.github.com/repos/ChadGoymer/githapi/git/refs",
+    url     = "https://api.github.com/repos/ChadGoymer/test-githapi/git/refs",
     type    = "POST",
-    payload = list(
-      ref = "refs/tags/zzz",
-      sha = "099944f501b2c2fba940f807b1028dbc5349f29c"))
+    payload = list(ref = test_tag, sha = master_sha))
   expect_is(created_tag, "list")
-  expect_identical(created_tag$ref, "refs/tags/zzz")
-  expect_identical(created_tag$object$sha, "099944f501b2c2fba940f807b1028dbc5349f29c")
+  expect_identical(created_tag$ref, test_tag)
+  expect_identical(created_tag$object$sha, master_sha)
 
-  expect_identical(attr(created_tag, "url"), "https://api.github.com/repos/ChadGoymer/githapi/git/refs")
+  expect_identical(attr(created_tag, "url"), "https://api.github.com/repos/ChadGoymer/test-githapi/git/refs")
   expect_identical(attr(created_tag, "request"), "POST")
   expect_identical(attr(created_tag, "status"), 201L)
   expect_true(length(attr(created_tag, "header")) > 1)
 
-  viewed_tag <- gh_request("https://api.github.com/repos/ChadGoymer/githapi/git/refs/tags/zzz", "GET")
+  viewed_tag <- gh_request(str_c("https://api.github.com/repos/ChadGoymer/test-githapi/git/", test_tag), "GET")
   expect_is(viewed_tag, "list")
-  expect_identical(viewed_tag$ref, "refs/tags/zzz")
-  expect_identical(viewed_tag$object$sha, "099944f501b2c2fba940f807b1028dbc5349f29c")
+  expect_identical(viewed_tag$ref, test_tag)
+  expect_identical(viewed_tag$object$sha, master_sha)
 
-  expect_identical(attr(viewed_tag, "url"), "https://api.github.com/repos/ChadGoymer/githapi/git/refs/tags/zzz")
+  expect_identical(attr(viewed_tag, "url"), str_c("https://api.github.com/repos/ChadGoymer/test-githapi/git/", test_tag))
   expect_identical(attr(viewed_tag, "request"), "GET")
   expect_identical(attr(viewed_tag, "status"), 200L)
   expect_true(length(attr(viewed_tag, "header")) > 1)
 
   updated_tag <- gh_request(
-    url     = "https://api.github.com/repos/ChadGoymer/githapi/git/refs/tags/zzz",
+    url     = str_c("https://api.github.com/repos/ChadGoymer/test-githapi/git/", test_tag),
     type    = "PATCH",
-    payload = list(sha = "ec311b3fa4ce294fadfc01dbfcfb35db3c88feda"))
+    payload = list(sha = test_sha))
   expect_is(updated_tag, "list")
-  expect_identical(updated_tag$ref, "refs/tags/zzz")
-  expect_identical(updated_tag$object$sha, "ec311b3fa4ce294fadfc01dbfcfb35db3c88feda")
+  expect_identical(updated_tag$ref, test_tag)
+  expect_identical(updated_tag$object$sha, test_sha)
 
-  expect_identical(attr(updated_tag, "url"), "https://api.github.com/repos/ChadGoymer/githapi/git/refs/tags/zzz")
+  expect_identical(attr(updated_tag, "url"), str_c("https://api.github.com/repos/ChadGoymer/test-githapi/git/", test_tag))
   expect_identical(attr(updated_tag, "request"), "PATCH")
   expect_identical(attr(updated_tag, "status"), 200L)
   expect_true(length(attr(updated_tag, "header")) > 1)
 
-  deleted_tag <- gh_request("https://api.github.com/repos/ChadGoymer/githapi/git/refs/tags/zzz", "DELETE")
+  deleted_tag <- gh_request(str_c("https://api.github.com/repos/ChadGoymer/test-githapi/git/", test_tag), "DELETE")
   expect_is(deleted_tag, "list")
 
-  expect_identical(attr(deleted_tag, "url"), "https://api.github.com/repos/ChadGoymer/githapi/git/refs/tags/zzz")
+  expect_identical(attr(deleted_tag, "url"), str_c("https://api.github.com/repos/ChadGoymer/test-githapi/git/", test_tag))
   expect_identical(attr(deleted_tag, "request"), "DELETE")
   expect_identical(attr(deleted_tag, "status"), 204L)
   expect_true(length(attr(deleted_tag, "header")) > 1)
 
-  expect_error(gh_request("https://api.github.com/repos/ChadGoymer/githapi/git/refs/tags/zzz", "GET"))
+  expect_error(gh_request(str_c("https://api.github.com/repos/ChadGoymer/test-githapi/git/", test_tag), "GET"))
 
 })